Fix build by checking lib, lib64, or other CMake lib directories
CMake sometimes builds in lib, or lib64, or lib32. It is different
on different systems. I do not know how CMake determines which lib
directory it will use.

In this commit we work around this problem by checking for the
existance of a few different lib directories and using whichever
is present.