Fix a heap corruption issue in get_monitor_info.
CString::from_raw takes ownership of a heap pointer; per the docs, "This should only ever be called with a pointer that was earlier obtained by calling into_raw on a CString." GetMonitorName returns a pointer to raylib-owned memory (specifically a glfw struct called _GLFWmonitor which has a "name" field). It ends up getting freed during raylib's CloseWindow.

Windows was yielding STATUS_HEAP_CORRUPTION for me after I started calling get_monitor_info, but it was pointing to __glfwFreeMonitor, which is (indirectly) called by CloseWindow.

I chose to copy the data into a String so the lifetime of the MonitorInfo is no longer tied to the lifetime of the GetMonitorName return value. It might be possible to instead change the lifetime of MonitorInfo to be tied to the RaylibHandle and then use a reference instead of a copy, but it seems like get_monitor_info is unlikely to be a hotspot and this preserves the current API.

I did not review other uses of CString::from_raw, since they are in modules I'm not currently using.