Replace renderloop with frame(), no longer expect mutable handles
While (im)mutable handles are fundamentally not a bad idea, the Rust semantics of mutable borrow mecanism doesn't really match what we are trying to do and can cause some weird errors. It is actually best to make things a lot simpler while losing a bit on guaranties on raylib state.
Additionally, replace the rather confusing "renderloop" with a simple frame callback mecanism.